requirements/local.txt
# 로컬 개발환경용 패키지
-r base.txt

# 개발 도구
django-debug-toolbar==4.2.0
ipython==8.17.2
django-extensions==3.2.3

# 테스트 도구
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
pytest==7.4.3
pytest-cov==4.1.0

# 코드 품질 도구
flake8==6.1.0
black==23.11.0
isort==5.12.0

# 로컬 개발 편의 도구
watchdog==3.0.0  # 파일 변경 감지
